+++ test/parallel/test-kill-segfault-freebsd.js
@@ -1,5 +1,10 @@
 'use strict';
 require('../common');
+
+// This test ensures Node.js doesn't crash on hitting Ctrl+C in order to
+// terminate the currently running process (especially on FreeBSD).
+// https://github.com/nodejs/node-v0.x-archive/issues/9326
+
 const assert = require('assert');
 const child_process = require('child_process');
