--- test/parallel/test-regress-GH-io-1068.js
@@ -1,3 +1,0 @@
-'use strict';
-require('../common');
-process.stdin.emit('end');
