+++ test/parallel/test-timers-setimmediate-infinite-loop.js
