+++ test/parallel/test-tls-tlswrap-segfault.js
@@ -1,18 +1,20 @@
 'use strict';
 const common = require('../common');
-
 if (!common.hasCrypto)
   common.skip('missing crypto');
+const fixtures = require('../common/fixtures');
+
+// This test ensures that Node.js doesn't incur a segfault while accessing
+// TLSWrap fields after the parent handle was destroyed.
+// https://github.com/nodejs/node/issues/5108
 
 const assert = require('assert');
 const tls = require('tls');
-const fixtures = require('../common/fixtures');
 
 const options = {
   key: fixtures.readKey('agent1-key.pem'),
   cert: fixtures.readKey('agent1-cert.pem')
 };
-
 
 const server = tls.createServer(options, function(s) {
   s.end('hello');
@@ -25,7 +27,6 @@ const server = tls.createServer(options, function(s) {
     putImmediate(client);
   });
 });
-
 
 function putImmediate(client) {
   setImmediate(function() {
